requirements.txt
asttokens==3.0.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
comm==0.2.3
dash==4.1.0
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
Flask==3.1.3
idna==3.11
importlib_metadata==9.0.0
ipykernel==7.2.0
ipython==9.10.1
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
narwhals==2.19.0
nest-asyncio==1.6.0
numpy==2.4.4
packaging==26.0
pandas==3.0.2
parso==0.8.6
platformdirs==4.9.6
plotly==6.7.0
prompt_toolkit==3.0.52
psutil==7.2.2
psycopg2-binary==2.9.11
pure_eval==0.2.3
Pygments==2.20.0
python-dateutil==2.9.0.post0
pyzmq==27.1.0
requests==2.33.1
retrying==1.4.2
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.5
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
wcwidth==0.6.0
Werkzeug==3.1.8
zipp==3.23.0
